chore: remove aboutSpouseForm custom component
