+++ libs/application/templates/financial-aid/src/fields/index.ts
@@ -13,7 +13,6 @@ export { default as Files } from './Files/Files'
 export { default as FileUploadContainer } from './FileUploadContainer/FileUploadContainer'
 export { default as TaxReturnFilesForm } from './TaxReturnFilesForm/TaxReturnFilesForm'
 export { default as Breakdown } from './Breakdown/Breakdown'
-export { default as AboutSpouseForm } from './AboutSpouseForm/AboutSpouseForm'
 export { default as PrivacyPolicyAccordion } from './PrivacyPolicyAccordion/PrivacyPolicyAccordion'
 export { default as ServiceCenter } from './ServiceCenter/ServiceCenter'
 export { default as ApplicantStatus } from './Status/ApplicantStatus'
